(DB) delete profil entity -> add attributes to user entity instead
