--- src/Entity/Profil.php
@@ -1,126 +1,0 @@
-<?php
-
-namespace App\Entity;
-
-use App\Repository\ProfilRepository;
-use Doctrine\DBAL\Types\Types;
-use Doctrine\ORM\Mapping as ORM;
-
-#[ORM\Entity(repositoryClass: ProfilRepository::class)]
-class Profil
-{
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column]
-    private ?int $id = null;
-
-    #[ORM\Column(length: 255)]
-    private ?string $nom = null;
-
-    #[ORM\Column(length: 255)]
-    private ?string $prenom = null;
-
-    #[ORM\Column(type: Types::DATE_MUTABLE)]
-    private ?\DateTimeInterface $ddn = null;
-
-    #[ORM\Column(length: 255, nullable: true)]
-    private ?string $tel = null;
-
-    #[ORM\Column(length: 255, nullable: true)]
-    private ?string $adr = null;
-
-    #[ORM\OneToOne(mappedBy: 'Profil', cascade: ['persist', 'remove'])]
-    private ?User $user = null;
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getNom(): ?string
-    {
-        return $this->nom;
-    }
-
-    public function setNom(string $nom): self
-    {
-        $this->nom = $nom;
-
-        return $this;
-    }
-
-    public function getPrenom(): ?string
-    {
-        return $this->prenom;
-    }
-
-    public function setPrenom(string $prenom): self
-    {
-        $this->prenom = $prenom;
-
-        return $this;
-    }
-
-    public function getDdn(): ?\DateTimeInterface
-    {
-        return $this->ddn;
-    }
-
-    public function setDdn(\DateTimeInterface $ddn): self
-    {
-        $this->ddn = $ddn;
-
-        return $this;
-    }
-
-    public function getTel(): ?string
-    {
-        return $this->tel;
-    }
-
-    public function setTel(?string $tel): self
-    {
-        $this->tel = $tel;
-
-        return $this;
-    }
-
-    public function getAdr(): ?string
-    {
-        return $this->adr;
-    }
-
-    public function setAdr(?string $adr): self
-    {
-        $this->adr = $adr;
-
-        return $this;
-    }
-
-    public function getUser(): ?User
-    {
-        return $this->user;
-    }
-
-    public function setUser(?User $user): self
-    {
-        // unset the owning side of the relation if necessary
-        if ($user === null && $this->user !== null) {
-            $this->user->setProfil(null);
-        }
-
-        // set the owning side of the relation if necessary
-        if ($user !== null && $user->getProfil() !== $this) {
-            $user->setProfil($this);
-        }
-
-        $this->user = $user;
-
-        return $this;
-    }
-
-    public function __toString(): string
-    {
-        return $this->nom;
-    }
-}
