--- src/Form/ProfilType.php
@@ -1,29 +1,0 @@
-<?php
-
-namespace App\Form;
-
-use App\Entity\Profil;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-class ProfilType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('nom')
-            ->add('prenom')
-            ->add('ddn')
-            ->add('tel')
-            ->add('adr')
-        ;
-    }
-
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver->setDefaults([
-            'data_class' => Profil::class,
-        ]);
-    }
-}
